3.14 move countFilms in utils.js
